gloo-timers: move the Javascript API bindings into a sys module
